+++ lizard_map/media/lizard_map/lizard_wms.js
@@ -125,7 +125,7 @@ function refreshLayers() {
 
 function showMap() {
     var options, base_layer, MapClickControl, MapHoverControl,
-        map_click_control, map_hover_control,
+        map_click_control,zoom_panel, map_hover_control,
         javascript_click_handler_name, javascript_hover_handler_name,
         $lizard_map_wms, projection, display_projection, start_extent,
         start_extent_left, start_extent_top, start_extent_right,
@@ -171,7 +171,8 @@ function showMap() {
             displayProjection: new OpenLayers.Projection(display_projection),  // "EPSG:4326"
             units: "m",
             numZoomLevels: 18,
-            maxExtent: max_extent
+            maxExtent: max_extent,
+	    controls: []
         };
     }
     else if (projection === "EPSG:28992")
@@ -181,7 +182,8 @@ function showMap() {
             displayProjection: new OpenLayers.Projection(display_projection),
             units: "m",
             resolutions: [364, 242, 161, 107, 71, 47, 31, 21, 14, 9, 6, 4, 2.7, 1.8, 0.9, 0.45, 0.2],
-            maxExtent: max_extent
+            maxExtent: max_extent,
+	    controls: []
         };
     }
     else
@@ -275,11 +277,43 @@ function showMap() {
         map_hover_control.activate();
     }
 
+    zoom_panel = new OpenLayers.Control.Panel();
+	 	zoom_panel.addControls([
+		new zoomSlider({zoomStopHeight:3}),
+				  ]);
+    map.addControl(zoom_panel);
+    map.addControl(new OpenLayers.Control.Navigation());
+
     // Zoom to startpoint. Important to parse numbers, else a bug in
     // OpenLayers will initially prevent "+" from working.
     map.setCenter(start_extent.getCenterLonLat(),
                   map.getZoomForExtent(start_extent));
 
+}
+
+
+function zoomSlider(options) {
+    this.control = new OpenLayers.Control.PanZoomBar(options);
+
+    OpenLayers.Util.extend(this.control,{
+        draw: function(px) {
+            // initialize our internal div
+            OpenLayers.Control.prototype.draw.apply(this, arguments);
+            px = this.position.clone();
+
+            // place the controls
+            this.buttons = [];
+
+            var sz = new OpenLayers.Size(18,18);
+            var centered = new OpenLayers.Pixel(px.x+sz.w/2, px.y);
+
+            this._addButton("zoomin", "zoom-plus-mini.png", centered.add(0, 5), sz);
+            centered = this._addZoomBar(centered.add(0, sz.h + 5));
+            this._addButton("zoomout", "zoom-minus-mini.png", centered, sz);
+            return this.div;
+        }
+    });
+    return this.control;
 }